+++ benchmark/simple_grid/gridspace_based.jl
@@ -12,8 +12,7 @@ function initialize_gridspace()
     N = grid_size[1]*grid_size[2]*grid_occupation
     for n in 1:N
         group = n < N / 2 ? 1 : 2
-        agent = GridSpaceAgent(n, (1, 1), group, false)
-        add_agent_single!(agent, model)
+        add_agent_single!(GridSpaceAgent, model, group, false)
     end
     return model
 end