+++ src/models/schelling.jl
@@ -20,8 +20,7 @@ function schelling(; numagents = 320, griddims = (20, 20), min_to_be_happy = 3)
     properties = Dict(:min_to_be_happy => min_to_be_happy)
     model = UnremovableABM(SchellingAgent, space; properties, scheduler = Schedulers.Randomly())
     for n in 1:numagents
-        agent = SchellingAgent(n, (1, 1), false, n < numagents / 2 ? 1 : 2)
-        add_agent_single!(agent, model)
+        add_agent_single!(SchellingAgent, model, false, n < numagents / 2 ? 1 : 2)
     end
     return model, schelling_agent_step!, dummystep
 end