+++ src/simulations/paramscan.jl
@@ -58,8 +58,7 @@ function initialize(; numagents = 320, griddims = (20, 20), min_to_be_happy = 3)
     model = ABM(SchellingAgent, space;
                 properties = properties, scheduler = Schedulers.randomly)
     for n in 1:numagents
-        agent = SchellingAgent(n, (1, 1), false, n < numagents / 2 ? 1 : 2)
-        add_agent_single!(agent, model)
+        add_agent_single!(SchellingAgent, model, n < numagents / 2 ? 1 : 2)
     end
     return model
 end